rwanda-travel-ai/requirements.txt
fastapi
uvicorn
pandas
joblib
requests
openpyxl
scikit-learn
python-dotenv
libretranslate
argostranslate
